+++ sample-code/cli-original/cookiecutter.json
@@ -3,8 +3,8 @@
     "email": "winnie.the.pooh@qbic.uni-tuebingen.de",
     "artifact_id": "sample-cli",
     "display_name": "Sample command-line Tool",
-    "main_class": "Sample",
+    "main_class": "SampleTool",
     "version": "1.0.0-SNAPSHOT",
-    "short_description": "Command-line utility to...",
+    "short_description": "openBIS connection tester",
     "copyright_holder": "QBiC"
 }
