+++ sample-code/cli-original/{{ cookiecutter.artifact_id }}/src/main/java/life/qbic/cli/{{ cookiecutter.main_class }}.java
