+++ sample-code/cli-original/{{ cookiecutter.artifact_id }}/src/main/resources/tool.properties
@@ -1,0 +1,3 @@
+# Properties set by maven at build time using "resource filtering"
+version=${project.version}
+project.url=${project.url}
