+++ sample-code/cli-original/{{ cookiecutter.artifact_id }}/src/test/resources/tool.properties
@@ -1,0 +1,3 @@
+# Properties set by maven at build time using "resource filtering"
+version=1.2.3-SNAPSHOT
+project.url=http://test.com
