+++ sample-code/cli/{{ cookiecutter.artifact_id }}/src/main/java/life/qbic/cli/{{ cookiecutter.main_class }}Command.java
@@ -1,0 +1,31 @@
+package life.qbic.cli;
+
+import picocli.CommandLine.Command;
+import picocli.CommandLine.Option;
+
+/**
+ * Abstraction of command-line arguments that will be passed to {@link {{ cookiecutter.main_class }}Tool} at construction time.
+ */
+@Command(
+   name="{{ cookiecutter.main_class }}",
+   description="{{ cookiecutter.short_description }}")
+public class {{ cookiecutter.main_class }}Command extends AbstractCommand {
+    // TODO: add your command-line options as members of this class using picocli's annotations, for instance:
+    //
+    // @Option(names={"-u", "--url"}, description="openBIS server URL.", required=true)
+    // String url;
+    //
+    // using package access level for these members will allow you access them within your main and test classes
+    //
+    // IMPORTANT: Typically you won't require a fancy constructor, but if you do, you must know that
+    //            ToolExecutor requires that all command classes contain a public constructor that takes no arguments.
+    //
+    //            If you need a custom constructor, make sure to provide a no-arguments public constructor as well.
+    //            See: https://docs.oracle.com/javase/tutorial/java/javaOO/constructors.html
+    @Option(names={"-u", "--url"}, description="openBIS server URL.", required=true)
+    private String url;
+    @Option(names={"-n", "--username"}, description="openBIS username.", required=true)
+    private String userName;
+    @Option(names={"-p", "--pass"}, description="openBIS password.", required=true)
+    private String password;
+}
