+++ sample-code/cli/{{ cookiecutter.artifact_id }}/src/main/resources/tool.properties
@@ -1,3 +1,4 @@
-# Properties set by maven at build time using "resource filtering"
-version=${project.version}
-project.url=${project.url}
+# These properties were set by maven at build time using "resource filtering"
+tool.version=${project.version}
+tool.repo.url=${project.url}
+tool.name=${project.artifactId}
